+++ lib/WorseReflection/Core/Inference/Resolver/MemberAccess/NodeContextFromMemberAccess.php
@@ -9,14 +9,19 @@ use Microsoft\PhpParser\Node\Expression\MemberAccessExpression;
 use Microsoft\PhpParser\Node\Expression\ScopedPropertyAccessExpression;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
 use Phpactor\WorseReflection\Core\Inference\Frame;
+use Phpactor\WorseReflection\Core\Inference\GenericMapResolver;
 use Phpactor\WorseReflection\Core\Inference\NodeContextFactory;
 use Phpactor\WorseReflection\Core\Inference\Symbol;
 use Phpactor\WorseReflection\Core\Inference\NodeContext;
-use Phpactor\WorseReflection\Core\Inference\GenericTypeResolver;
+use Phpactor\WorseReflection\Core\Reflection\Collection\ReflectionInterfaceCollection;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionClass;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionClassLike;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionInterface;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionEnum;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionMember;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionProperty;
 use Phpactor\WorseReflection\Core\Type\ClassType;
+use Phpactor\WorseReflection\Core\Type\GenericClassType;
 use Phpactor\WorseReflection\Core\Type\GlobbedConstantUnionType;
 use Phpactor\WorseReflection\Core\Type\SelfType;
 use Phpactor\WorseReflection\Core\Type\StaticType;
@@ -30,11 +35,11 @@ use Phpactor\WorseReflection\Core\Inference\NodeContextResolver;
 
 class NodeContextFromMemberAccess
 {
-    private GenericTypeResolver $genericResolver;
+    private GenericMapResolver $resolver;
 
-    public function __construct()
+    public function __construct(GenericMapResolver $resolver)
     {
-        $this->genericResolver = new GenericTypeResolver();
+        $this->resolver = $resolver;
     }
 
     public function infoFromMemberAccess(NodeContextResolver $resolver, Frame $frame, Type $classType, Node $node): NodeContext
@@ -127,16 +132,29 @@ class NodeContextFromMemberAccess
         $inferredType = $member->inferredType();
 
         if ($member instanceof ReflectionProperty) {
-            $inferredType = self::getFrameTypesForPropertyAtPosition(
+            $propertyType = self::getFrameTypesForPropertyAtPosition(
                 $frame,
                 $member->name(),
                 $subType,
                 $node->getEndPosition(),
             );
+            if ($propertyType) {
+                $inferredType = $propertyType;
+            }
         }
 
 
-        $inferredType = $this->genericResolver->resolveMemberType($subType, $member, $inferredType);
+        $declaringClass = self::declaringClass($member);
+
+        if (count($declaringClass->templateMap())) {
+            $templateMap = $this->resolver->resolveClassTemplateMap($subType, $declaringClass->name(), $subType instanceof GenericClassType ? $subType->arguments() : []);
+            $inferredType = $inferredType->map(function (Type $type) use ($templateMap): Type {
+                if ($templateMap && $templateMap->has($type->short())) {
+                    return $templateMap->get($type->short());
+                }
+                return $type;
+            });
+        }
 
         // unwrap static and self types (including $this which extends Static) and any nested globbed constant unions
         $inferredType = $inferredType->map(function (Type $type) {
@@ -177,5 +195,37 @@ class NodeContextFromMemberAccess
         }
 
         return $variable->type();
+    }
+
+    private static function declaringClass(ReflectionMember $member): ReflectionClassLike
+    {
+        $reflectionClass = $member->declaringClass();
+
+        if (!$reflectionClass instanceof ReflectionClass) {
+            return $reflectionClass;
+        }
+
+        $interface = self::searchInterfaces($reflectionClass->interfaces(), $member->name());
+
+        if (!$interface) {
+            return $reflectionClass;
+        }
+
+        return $interface;
+    }
+
+    private static function searchInterfaces(ReflectionInterfaceCollection $collection, string $memberName): ?ReflectionInterface
+    {
+        foreach ($collection as $interface) {
+            if ($interface->methods()->has($memberName)) {
+                return $interface;
+            }
+
+            if (null !== $interface = self::searchInterfaces($interface->parents(), $memberName)) {
+                return $interface;
+            }
+        }
+
+        return null;
     }
 }