+++ lib/WorseReflection/Core/Type/GenericClassType.php
@@ -39,12 +39,12 @@ class GenericClassType extends ReflectedClassType implements IterableType, Class
      */
     public function arguments(): array
     {
-        return $this->arguments;
+        return array_values($this->arguments);
     }
 
     public function iterableValueType(): Type
     {
-        return IterableTypeResolver::resolveIterable($this, $this->arguments);
+        return IterableTypeResolver::resolveIterable($this->reflector, $this, $this->arguments);
     }
 
     public function toPhpString(): string