+++ lib/WorseReflection/Core/Type/ReflectedClassType.php
@@ -112,10 +112,10 @@ class ReflectedClassType extends ClassType
 
         foreach ($genericTypes as $genericType) {
             if (!$genericType instanceof GenericClassType) {
-                return new MissingType();
+                continue;
             }
 
-            $type = IterableTypeResolver::resolveIterable($genericType, $genericType->arguments());
+            $type = IterableTypeResolver::resolveIterable($this->reflector, $genericType, $genericType->arguments());
             if (!$type->isDefined()) {
                 continue;
             }